+++ toys-core/CHANGELOG.md
@@ -6,6 +6,7 @@ Now declaring this alpha quality, and soft-locking the usage. Will try not to
 make further backward-incompatible changes.
 
 * CHANGED: Utils::Terminal#confirm default is now unset by default
+* IMPROVED: Toys::Utils::Gems can suppress the confirmation prompt
 * IMPROVED: Toys files can now have magic comments.
 
 ### 0.3.11 / 2018-07-02